## Description

Resolves CodeQL issues.

Another lighter pass of CodeQL will occur shortly in the future.

- [ ] Breaking change?
  - Will this change break pre-existing builds or functionality without action being taken?
  **No**

## How This Was Tested

- Verified CI build
- Checked before and after CodeQL results
- Verified boot on QemuQ35Pkg

## Integration Instructions

N/A - If CodeQL was run before these changes, run again to get latest results

+++ MsWheaPkg/Library/MsWheaEarlyStorageLib/MsWheaEarlyStorageLib.c
@@ -376,7 +376,7 @@ MsWheaESCalculateChecksum16 (
   )
 {
   UINT16      Data;
-  UINT8       Index;
+  UINT32      Index;
   UINT16      Sum;
   EFI_STATUS  Status;
 
@@ -401,7 +401,7 @@ MsWheaESCalculateChecksum16 (
   Header->Checksum = *Checksum;
 
   for (Index = 0; Index < Header->ActiveRange; Index += sizeof (Data)) {
-    Status = MsWheaEarlyStorageRead (&Data, sizeof (Data), MS_WHEA_EARLY_STORAGE_DATA_OFFSET + Index);
+    Status = MsWheaEarlyStorageRead (&Data, (UINT8)sizeof (Data), (UINT8)(MS_WHEA_EARLY_STORAGE_DATA_OFFSET + Index));
     if (EFI_ERROR (Status) != FALSE) {
       DEBUG ((DEBUG_ERROR, "%a: Reading Early Storage %d failed %r\n", __FUNCTION__, Index, Status));
       goto Cleanup;
